+++ api/main.go
@@ -54,7 +54,7 @@ func runHttpServer(server *Server) error {
 	// it's better to use websockets protocols for streaming and interactive one-off commands
 	mux.Handle("/ws/v1/logs", ws("appLogs", server.LogStreamWs))
 	mux.Handle("/ws/v1/exec", ws("processExec", server.ProcessRunWs))
-
+	mux.Handle("/ws/v1/proxy", ws("appProxy", server.ResourceProxy))
 	mux.Handle("/", gwmux)
 
 	fmt.Printf("Starting server on http=%d and grpc=%d ports\n", port, rpcPort)